requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-google-genai
faiss-cpu
pydantic[email]
PyMuPDF
langchain-community
bcrypt
